+++ tests/codeception/unit/CropActionTest.php
@@ -2,7 +2,7 @@
 
 namespace tests\codeception\unit;
 
-/**
+/*
  *          _)             __|  | _)
  * \ \  \ / | (_-<   -_) __ \  |  |    \
  *  \_/\_/ _| ___/ \___| ___/ _| _| _| _|
@@ -13,12 +13,17 @@ namespace tests\codeception\unit;
 
 use Yii;
 use yii\helpers\Json;
-use Codeception\Specify;
 
 class CropActionTest extends TestCase
 {
-    use Specify;
+    /**
+     * @var \frontend\tests\UnitTester
+     */
+    protected $tester;
 
+    /**
+     * {@inheritdoc}
+     */
     protected function _after()
     {
         if (is_file('tests/codeception/unit/data/img/cropped/img.jpeg')) {
@@ -27,8 +32,12 @@ class CropActionTest extends TestCase
     }
 
     /**
+     * Тест на незаполненный атрибут 'path'.
+     *
      * @expectedException        yii\base\InvalidConfigException
-     * @expectedExceptionMessage Атрибут "path" не может быть пустым
+     * @expectedExceptionMessage Атрибут "path" пуст или не является строкой.
+     *
+     * @method testEmptyPath
      */
     public function testEmptyPath()
     {
@@ -36,8 +45,12 @@ class CropActionTest extends TestCase
     }
 
     /**
+     * Тест на незаполненный атрибут 'url'.
+     *
      * @expectedException        yii\base\InvalidConfigException
-     * @expectedExceptionMessage Атрибут "url" не может быть пустым
+     * @expectedExceptionMessage Атрибут "url" пуст или не является строкой.
+     *
+     * @method testEmptyUrl
      */
     public function testEmptyUrl()
     {
@@ -45,9 +58,13 @@ class CropActionTest extends TestCase
     }
 
     /**
+     * Тест на соответствие модели экземпляру класса 'yii\db\BaseActiveRecord'.
+     *
      * @expectedException        yii\base\InvalidConfigException
      * @expectedExceptionMessage Атрибут "model" не является экземпляром
      *                           класса "yii\db\BaseActiveRecord"
+     *
+     * @method testModelInstanceofClass
      */
     public function testModelInstanceofClass()
     {
@@ -55,20 +72,25 @@ class CropActionTest extends TestCase
         Yii::$app->runAction('test/crop-model-instanceof-class');
     }
 
+    /**
+     * Тест на удачную обработку изображения.
+     *
+     * @method testCropImage
+     */
     public function testCropImage()
     {
         Yii::$app->getSession()->set('tempImage', 'img.jpeg');
 
         $_SERVER['REQUEST_METHOD'] = 'POST';
         $_POST = [
-            'imgUrl' => parent::IMG_DIR . '/img.jpeg',
+            'imgUrl' => parent::IMG_DIR.'/img.jpeg',
             'imgW' => 100,
             'imgH' => 100,
             'imgY1' => 25,
             'imgX1' => 25,
             'cropW' => 50,
             'cropH' => 50,
-            'rotation' => 0
+            'rotation' => 0,
         ];
 
         $json = [
@@ -76,7 +98,6 @@ class CropActionTest extends TestCase
             'url' => '/img/cropped/img.jpeg',
         ];
 
-        expect('Изображение успешно сохранено', Yii::$app->runAction('test/crop'))
-            ->equals(Json::encode($json));
+        $this->tester->assertEquals(Json::encode($json), Yii::$app->runAction('test/crop'));
     }
 }