+++ tests/codeception/unit/UploadActionTest.php
@@ -2,7 +2,7 @@
 
 namespace tests\codeception\unit;
 
-/**
+/*
  *          _)             __|  | _)
  * \ \  \ / | (_-<   -_) __ \  |  |    \
  *  \_/\_/ _| ___/ \___| ___/ _| _| _| _|
@@ -13,14 +13,19 @@ namespace tests\codeception\unit;
 
 use Yii;
 use yii\helpers\Json;
-use Codeception\Specify;
 use yii\helpers\FileHelper;
 use org\bovigo\vfs\vfsStream;
 
 class UploadActionTest extends TestCase
 {
-    use Specify;
+    /**
+     * @var \frontend\tests\UnitTester
+     */
+    protected $tester;
 
+    /**
+     * {@inheritdoc}
+     */
     protected function _after()
     {
         unset($_SERVER['REQUEST_METHOD']);
@@ -28,8 +33,12 @@ class UploadActionTest extends TestCase
     }
 
     /**
+     * Тест на незаполненный атрибут 'tempPath'.
+     *
      * @expectedException        yii\base\InvalidConfigException
-     * @expectedExceptionMessage Атрибут "tempPath" не может быть пустым
+     * @expectedExceptionMessage Атрибут "tempPath" пуст или не является строкой.
+     *
+     * @method testEmptyPath
      */
     public function testEmptyPath()
     {
@@ -37,8 +46,12 @@ class UploadActionTest extends TestCase
     }
 
     /**
+     * Тест на незаполненный атрибут 'tempUrl'.
+     *
      * @expectedException        yii\base\InvalidConfigException
-     * @expectedExceptionMessage Атрибут "tempUrl" не может быть пустым
+     * @expectedExceptionMessage Атрибут "tempUrl" пуст или не является строкой.
+     *
+     * @method testEmptyUrl
      */
     public function testEmptyUrl()
     {
@@ -46,9 +59,13 @@ class UploadActionTest extends TestCase
     }
 
     /**
+     * Тест на соответствие модели экземпляру класса 'yii\base\Model'.
+     *
      * @expectedException        yii\base\InvalidConfigException
      * @expectedExceptionMessage Атрибут "model" не является экземпляром
      *                           класса "yii\base\Model"
+     *
+     * @method testModelInstanceofClass
      */
     public function testModelInstanceofClass()
     {
@@ -56,9 +73,14 @@ class UploadActionTest extends TestCase
         Yii::$app->runAction('test/upload-model-instanceof-class');
     }
 
-    public function testUploadImage()
+    /**
+     * Тест на неудачную загрузку изображения.
+     *
+     * @method testNotSuccessfulUploadImage
+     */
+    public function testNotSuccessfulUploadImage()
     {
-        $path = vfsStream::url(parent::ROOT_DIR . '/' . parent::IMG_DIR . '/img.jpeg');
+        $path = vfsStream::url(parent::ROOT_DIR.'/'.parent::IMG_DIR.'/img.jpeg');
 
         $_SERVER['REQUEST_METHOD'] = 'POST';
         $_FILES = [
@@ -67,30 +89,45 @@ class UploadActionTest extends TestCase
                 'type' => FileHelper::getMimeType($path),
                 'size' => filesize($path),
                 'tmp_name' => $path,
-                'error' => UPLOAD_ERR_OK
-            ]
+                'error' => UPLOAD_ERR_OK,
+            ],
         ];
 
-        $this->specify('Неуспешная загрузка изображения (неверное расширение файла)', function () {
-            $json = [
-                'status' => 'error',
-                'message' => 'Разрешена загрузка файлов только со следующими расширениями: png.',
-            ];
+        $json = [
+            'status' => 'error',
+            'message' => 'Разрешена загрузка файлов только со следующими расширениями: png.',
+        ];
 
-            expect('Не удалось сохранить изображение', Yii::$app->runAction('test/upload-error'))
-                ->equals(Json::encode($json));
-        });
+        $this->tester->assertEquals(Json::encode($json), Yii::$app->runAction('test/upload-error'));
+    }
 
-        $this->specify('Успешная загрузка изображения', function () {
-            $json = [
-                'status' => 'success',
-                'url' => '/img/temp/img.jpeg',
-                'width' => 100,
-                'height' => 100
-            ];
+    /**
+     * Тест на удачную загрузку изображения.
+     *
+     * @method testSuccessfulUploadImage
+     */
+    public function testSuccessfulUploadImage()
+    {
+        $path = vfsStream::url(parent::ROOT_DIR.'/'.parent::IMG_DIR.'/img.jpeg');
 
-            expect('Изображение успешно сохранено', Yii::$app->runAction('test/upload'))
-                ->equals(Json::encode($json));
-        });
+        $_SERVER['REQUEST_METHOD'] = 'POST';
+        $_FILES = [
+            'img' => [
+                'name' => 'img.jpeg',
+                'type' => FileHelper::getMimeType($path),
+                'size' => filesize($path),
+                'tmp_name' => $path,
+                'error' => UPLOAD_ERR_OK,
+            ],
+        ];
+
+        $json = [
+            'status' => 'success',
+            'url' => '/img/temp/img.jpeg',
+            'width' => 100,
+            'height' => 100,
+        ];
+
+        $this->tester->assertEquals(Json::encode($json), Yii::$app->runAction('test/upload'));
     }
 }